My Route alerts now displays map with alerts
