+++ wsdot/MyRouteCell.swift
@@ -24,7 +24,6 @@ class MyRouteCell: UITableViewCell {
     @IBOutlet weak var titleLabel: UILabel!
     @IBOutlet weak var setButton: UIButton!
     @IBOutlet weak var checkAlertsButton: UIButton!
-    @IBOutlet weak var checkMapButton: UIButton!
     @IBOutlet weak var settingsButton: UIButton!
 
 }
